CAUSEWAY-3616: moves background job control from the job and into a singleton service
